Update category ID
Change the property _id to id.
Update the logic for building category options.
