+++ src/components/category-list/CategoryList.tsx
@@ -1,7 +1,8 @@
 import { ResponseData } from '@/utils/data-fetching-utils';
 import styles from './CategoryList.module.css';
 import ProductCategory from './product-category/ProductCategory';
-import { Category, getCategories } from '@/api-services/market-service';
+import { getCategories } from '@/api-services/market-service';
+import { Category } from '@/model/market-data-model';
 import { log } from 'console';
 
 export default async function CategoryList() {
@@ -18,7 +19,7 @@ export default async function CategoryList() {
             {
                 categories.data.map(
                     category =>
-                        <ProductCategory key={category._id} imageSource={category.imageUrl} categoryName={category.name} />
+                        <ProductCategory key={category.id} imageSource={category.imageUrl} categoryName={category.name} />
                 )
             };
         </div>
